fix bug and reform code
