+++ extension-impl/tracer-opentracing/src/main/java/com/alipay/sofa/rpc/tracer/sofatracer/log/digest/RpcServerDigestSpanJsonEncoder.java
@@ -42,6 +42,7 @@ public class RpcServerDigestSpanJsonEncoder implements SpanEncoder<SofaTracerSpa
         jsb.appendBegin("timestamp", Timestamp.format(span.getEndTime()));
         //添加其他字段
         this.appendSlot(jsb, span);
+        jsb.appendEnd();
         return jsb.toString();
     }
