+++ extension-impl/tracer-opentracing/src/main/java/com/alipay/sofa/rpc/tracer/sofatracer/log/stat/RpcClientStatJsonReporter.java
@@ -16,12 +16,6 @@
  */
 package com.alipay.sofa.rpc.tracer.sofatracer.log.stat;
 
-import com.alipay.common.tracer.core.appender.builder.JsonStringBuilder;
-import com.alipay.common.tracer.core.appender.file.LoadTestAwareAppender;
-import com.alipay.common.tracer.core.appender.self.SelfLog;
-import com.alipay.common.tracer.core.appender.self.Timestamp;
-import com.alipay.common.tracer.core.reporter.stat.model.StatKey;
-import com.alipay.common.tracer.core.reporter.stat.model.StatMapKey;
 import com.alipay.sofa.rpc.tracer.sofatracer.log.tags.RpcSpanTags;
 
 import java.util.Map;
@@ -32,8 +26,6 @@ import java.util.Map;
  * @author <a href=mailto:guanchao.ygc@antfin.com>GuanChao Yang</a>
  */
 public class RpcClientStatJsonReporter extends AbstractRpcStatJsonReporter {
-
-    private static JsonStringBuilder buffer = new JsonStringBuilder();
 
     /**
      * Instantiates a new Rpc client stat json reporter.
@@ -62,43 +54,4 @@ public class RpcClientStatJsonReporter extends AbstractRpcStatJsonReporter {
         return tagsWithStr.get(RpcSpanTags.REMOTE_ZONE);
     }
 
-    @Override
-    public void print(StatKey statKey, long[] values) {
-        if (this.isClosePrint.get()) {
-            //关闭统计日志输出
-            return;
-        }
-
-        StatMapKey statMapKey = (StatMapKey) statKey;
-
-        buffer.reset();
-        buffer.appendBegin("time", Timestamp.currentTime());
-        buffer.append("stat.key", this.statKeySplit(statMapKey));
-        buffer.append("count", values[0]);
-        buffer.append("total.cost.milliseconds", values[1]);
-        buffer.append("success", statMapKey.getResult());
-        buffer.appendEnd();
-        try {
-            if (appender instanceof LoadTestAwareAppender) {
-                ((LoadTestAwareAppender) appender).append(buffer.toString(), statKey.isLoadTest());
-            } else {
-                appender.append(buffer.toString());
-            }
-            // 这里强制刷一次
-            appender.flush();
-        } catch (Throwable t) {
-            SelfLog.error("统计日志<" + statTracerName + ">输出异常", t);
-        }
-    }
-
-    private String statKeySplit(StatMapKey statKey) {
-        JsonStringBuilder jsonBufferKey = new JsonStringBuilder();
-        Map<String, String> keyMap = statKey.getKeyMap();
-        jsonBufferKey.appendBegin();
-        for (Map.Entry<String, String> entry : keyMap.entrySet()) {
-            jsonBufferKey.append(entry.getKey(), entry.getValue());
-        }
-        jsonBufferKey.appendEnd(false);
-        return jsonBufferKey.toString();
-    }
 }
